TASK/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
bcrypt==4.1.2
supabase==1.0.0
